move related files together
