+++ frontend/src/modules/Datasets/components/index.js
@@ -1,0 +1,2 @@
+export * from './DatasetAWSActions';
+export * from './DatasetConsoleAccess';
